requirements.txt
apache-beam[gcp]==2.63.0
apache-airflow[google]==2.10.5